fixed graphs and trees to use value instead of key and value
Previously, the graph classes required explicitly setting a key and a
value. But the only place the key is actually required is to hash the
graph vertices during stateful search like in dfs and bfs. So I made
this explicit with a new auto generated __hashKey__ which can be used to
hash the objects like this: dict[vertex.toHashKey()] = vertex. For
comparisons in binary search the value can be used directly. A custom
compare function can be provided as an option to the binary search tree
class to compare custom objects. For regular strings and numbers the
default comparator function (which returns 1, 0, -1 based on >, ==, <)
will work.